# perf(motion_velocity_planner): resample trajectory and build rtree with packing algorithm