## Description

<!-- Write a brief description of this PR. -->
Cherry-pick of https://github.com/autowarefoundation/autoware.universe/pull/7730 https://github.com/autowarefoundation/autoware.universe/pull/7732 #7683 and #7633

## Tests performed
Psim and perception replayer.

Performance of the `out_of_lane` (most heavy module) on a rosbag with many dynamic objects:
- Without the PR: ~50ms runtime with peaks at 80ms.
![image](https://github.com/tier4/autoware.universe/assets/78338830/2fb772e2-1398-4cb1-adfe-2b6a5e02ddab)

- With the PR: ~15ms runtime with peaks at 25ms
![image](https://github.com/tier4/autoware.universe/assets/78338830/7b4edcaa-6eee-4b6f-8790-f60072d82a56)


<!-- Describe how you have tested this PR. -->
<!-- Although the default value is set to "Not Applicable.", please update this section if the type is either [feat, fix, perf], or if requested by the reviewers. -->

Not applicable.

## Effects on system behavior

<!-- Describe how this PR affects the system behavior. -->

Not applicable.

## Interface changes

<!-- Describe any changed interfaces, such as topics, services, or parameters, including debugging interfaces -->

## Pre-review checklist for the PR author

The PR author **must** check the checkboxes below when creating the PR.

- [ ] I've confirmed the [contribution guidelines].
- [ ] The PR follows the [pull request guidelines].

## In-review checklist for the PR reviewers

The PR reviewers **must** check the checkboxes below before approval.

- [ ] The PR follows the [pull request guidelines].

## Post-review checklist for the PR author

The PR author **must** check the checkboxes below before merging.

- [ ] There are no open discussions or they are tracked via tickets.

After all checkboxes are checked, anyone who has write access can merge the PR.

[contribution guidelines]: https://autowarefoundation.github.io/autoware-documentation/main/contributing/
[pull request guidelines]: https://autowarefoundation.github.io/autoware-documentation/main/contributing/pull-request-guidelines/